Added timestamps to chat messages  (#266)
* Added timestamps to chat messages

* Added toggle for them

Currently does a little funkiness. In order for the update to happen in realtime I have to discard all the chat messages in chat and re-list them. Unfortunately the assign for chat_messages is blank. So it runs the list_chat_messages function and retrieves that output.

* Made toggling timestamps less "intrusive"

Also added a limit option to the Chat.list_chat_messages function for potential use in the future and now.

* Added fallback to UTC if can't convert

* Added checks for logged out viewers

* First batch of tests

* Added tests for the chat liveview component

* Change IDs to classes

* Removed ? from show_timestamps

* Added a default for the list_chat_message on limit

* Clean up the multi chat messages function

* Move timestamp event handler to fix build warning

* Make toggle timestamps less "heavy"

* Add back the temporary assign

* Moved the timestamp display to JS

* Changed timestamp test

* Moved timestamp setting over to user_preference

* Small fixes

Fixed weird message cutoffs when toggling on timestamps if the timestamp overflows onto the second line.
Toggling now updates all previous messages